WARNING! Fixing explicit parameter types in classes. WARNING! Currently only Pure Safe params are available in classes, so the following code would not work: class Vector a:     x :: a
def main self:
    a = readIntFromFile()
    v = Vector 0
    v.x = a # a is of type IO(s(Int))

but it could be fixed by introducing keyword eval.
It is not implemented yet, but current pre-implementation always outputs
IO value, which combined with top level function assignment moves the IO
onto the function itself and results in Pure variable value.
So the abouve code would be fixed by:

    v.x = eval a # now main is IO (...) and v.x is Pure(s(Int))

assuming, that s is safe!