Overhauled nix expressions
Still not the best, because it doesn't call gencabal before building and
has ugly relative paths in each nix expression. However, it's almost
self-contained, so all relevant files are in the repo.

The only thing that should be enabled in home directory is to allow
unfree packages to be built.