Somebody forgot to NEVER commit case-sens. files
I mean posting both 'tmp.hs' and 'Tmp.hs'. This is a BIG problem on
Windows. This is a problem on OS X (and I can't set the HFS+ filesystem
to case-sensitive mode, photoshop and other Adobe-crap will stop
working. Yes, really).

If I could post a git-hook to reject such commits I would do that. I
can't "because github".